Tools: Renamed unit-tests to tests and moved runtime to tools.
