+++ package.json
@@ -58,10 +58,10 @@
         "scripts:bin": "rm -rf bin/ ; tsc -p src/CLI/ ; chmod +x bin/*.js",
         "scripts:code": "rm -rf code/ ; tsc -p src/",
         "reset": "rm -rf bin/ build/ code/ node_modules/ ; npm i",
-        "test": "npm run webpack && npm run test:eslint && npm run test:unit-tests",
+        "test": "npm run webpack && npm run test:eslint && npm run test:units",
         "test:eslint": "eslint .",
         "test:morningstar": "npx -y newman run --bail tmp/Collection.json -e tmp/Environment.json",
-        "test:unit-tests": "ts-node test/unit-tests/runtime",
+        "test:units": "ts-node tools/tests",
         "watch": "tsc -p src/ -w",
         "webpack": "npm run scripts && webpack && npm run webpack:dts",
         "webpack:dts": "npm run webpack:dts:cm && npm run webpack:dts:dbm && npm run webpack:dts:dgm && npm run webpack:dts:hcm",
