+++ tests/TimeSeries/Rating.test.ts
@@ -1,19 +1,19 @@
 import * as Assert from 'node:assert/strict';
-import * as MC from '../../../code/connectors-morningstar.src';
+import * as MC from '../../code/connectors-morningstar.src';
 
-export async function growthLoad (
+export async function ratingLoad (
     api: MC.Shared.MorningstarAPIOptions
 ) {
     const connector = new MC.TimeSeriesConnector({
         api,
         currencyId: 'EUR',
-        endDate: '2020-01-31',
+        endDate: '2020-12-31',
         securities: [{
             id: 'F0GBR04S23',
             idType: 'MSID'
         }],
         series: {
-            type: 'Growth'
+            type: 'Rating'
         },
         startDate: '2020-01-01'
     });
@@ -25,8 +25,8 @@ export async function growthLoad (
 
     Assert.ok(
         connector.converter instanceof
-        MC.TimeSeriesConverters.GrowthSeriesConverter,
-        'Converter should be instance of TimeSeries GrowthSeriesConverter.'
+        MC.TimeSeriesConverters.RatingSeriesConverter,
+        'Converter should be instance of TimeSeries RatingSeriesConverter.'
     );
 
     await connector.load();
@@ -39,8 +39,8 @@ export async function growthLoad (
 
     Assert.strictEqual(
         connector.table.getRowCount(),
-        31,
-        'Connector table should have 31 growth rows.'
+        12,
+        'Connector table should have 12 expected rating rows.'
     );
 
 }
